+++ cloveretl.engine/src/org/jetel/graph/TransformationGraphXMLReaderWriter.java
@@ -352,7 +352,7 @@ public class TransformationGraphXMLReaderWriter {
 	        //get debug mode
 	        graph.setDebugMode(grfAttributes.getString("debugMode", "true"));
 	        //get debugMaxRecords
-	        graph.setDebugMaxRecords(grfAttributes.getInteger("debugMaxRecords", 0));
+	        graph.setDebugMaxRecords(grfAttributes.getLong("debugMaxRecords", 0));
 	        
 	        graph.setAuthor(grfAttributes.getString(AUTHOR_ATTRIBUTE, null));
 	        graph.setRevision(grfAttributes.getString(REVISION_ATTRIBUTE, null));
@@ -559,7 +559,7 @@ public class TransformationGraphXMLReaderWriter {
 	 * @param  nodes         Description of Parameter
 	 * @since                May 24, 2002
 	 */
-	private void instantiateEdges(NodeList edgeElements, Map<String, Object> metadata, boolean graphDebugMode, int graphDebugMaxRecords) throws XMLConfigurationException,GraphConfigurationException {
+	private void instantiateEdges(NodeList edgeElements, Map<String, Object> metadata, boolean graphDebugMode, long graphDebugMaxRecords) throws XMLConfigurationException,GraphConfigurationException {
 		String edgeID="unknown";
 		String edgeMetadataID;
 		String fromNodeAttr;
@@ -567,7 +567,7 @@ public class TransformationGraphXMLReaderWriter {
 		String edgeType = null;
         boolean debugMode = false;
         String debugFilterExpression = null;
-        int debugMaxRecords = 0;
+        long debugMaxRecords = 0;
         boolean debugLastRecords = true;
         boolean debugSampleData = false;
         boolean fastPropagate = false;
@@ -595,7 +595,7 @@ public class TransformationGraphXMLReaderWriter {
 			debugMode = attributes.getBoolean("debugMode", false);
             
             if (graphDebugMaxRecords == 0) { // if this value isn't defined for whole graph 
-            	debugMaxRecords = attributes.getInteger("debugMaxRecords", 0);
+            	debugMaxRecords = attributes.getLong("debugMaxRecords", 0);
             } else {
             	debugMaxRecords = graphDebugMaxRecords;
             }
